Combine manage and store
Looks /wonderful/ now.
